+++ packages/@react-spectrum/actiongroup/src/index.ts
@@ -11,3 +11,4 @@
  */
 
 export * from './ActionGroup';
+export {Item} from '@react-stately/collections';
